+++ custodian/__init__.py
@@ -4,6 +4,6 @@ framework written in Python.
 """
 
 __author__ = "Shyue Ping Ong, William Davidson Richards, Stephen Dacek, Xiaohui Qu, Matthew Horton, Samuel M. Blau"
-__version__ = "2021.2.8"
+__version__ = "2022.1.17"
 
 from .custodian import Custodian